+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.4.0
+
+Cosmetic changes to better mimic Datomic API. Useful for sharing code between Datomic and DataScript:
+
+- Added `tempid`, `resolve-tempid`, `db`, `transact`, `transact-async`, `index-range`, `squuid`, `squuid-time-millis`
+- [ BREAKING ] renamed `transact` to `with`, `with` to `db-with`
+
 # 0.3.1
 
 - Optimized speed of DB’s `equiv` and `hash`, Datom’s `hash`
@@ -9,7 +16,7 @@
 Proper entities implementation:
 
 - Entities are now lazy and implement usual Map protocols
-- When accessing attribute of `:db/valueType :db.type/ref`, its value will be automatically expanded to entites, allowing for recursive exploration of entities graphs (e.g. `(-> (d/entity db 42) :parent :parent :children)`)
+- [ BREAKING ] When accessing attribute of `:db/valueType :db.type/ref`, its value will be automatically expanded to entites, allowing for recursive exploration of entities graphs (e.g. `(-> (d/entity db 42) :parent :parent :children)`)
 - Entities support backwards navigation (e.g. `(:person/_friends (d/entity db 42))`)
 
 # 0.2.1